refactor(datasets): извлечён DataSetsResource + системные наборы на нём (#210 ось видимости, PR2)
Второй шаг унификации ресурсов оси scope. Извлечён `DataSetsResource({scope,scopeId})` —
браузер наборов данных (загрузка + список файлов с источниками через SourcesExpander),
scope-параметризованный, БЕЗ чипа области. `DataSetsPage` сведён к тонкой обёртке (+ «Шаблоны
обработки»). Канонический FileRow (System-размер) с scope-параметрами. Задел под detail-панели
scope-страниц. ScopedDataSetsPanel мигрируем при постройке scope-страниц.
